Use human-readable amounts, add KMS address script, remove dead CloudWatch metrics
## Amount handling

All invoice amounts are now specified in human-readable units rather than
native blockchain units:
- Native SUI: `0.1` means 0.1 SUI (not 100,000,000 MIST)
- Token payments: `0.01` means 0.01 USDC (not 10,000 micro-USDC)

Changes:
- `shared/src/models.rs`: `Invoice.amount` changed from `u64` to `f64`
- `invoice-generator/src/main.rs`: accepts `amount: f64` from API request;
  stores the human-readable value in DynamoDB as-is
- `watcher/src/main.rs`: reads human-readable amount and converts to native
  units before comparing against on-chain balance (× 1e9 for SUI,
  × 10^token_decimals for tokens)
- `sweeper/src/main.rs`: converts human-readable SUI amount to MIST
  (`amount_mist`) before passing to `transfer_sui`; removes hardcoded
  "MIST" label from SNS alert message
- `infra/stack.ts`: API Gateway amount validation changed from
  `INTEGER` with a MIST-based maximum to `NUMBER` with `exclusiveMinimum: 0`
- `README.md`: all API examples updated to human-readable amounts

## KMS hot wallet address script

Token sweeps use sponsored transactions where the KMS Ed25519 key pays gas.
Operators need to fund this wallet before processing token invoices, but
previously there was no tooling to derive its SUI address.

- `scripts/get-kms-address.js`: new script that queries CloudFormation for
  the `KmsKeyId` stack output, fetches the DER-encoded public key from KMS,
  extracts the raw 32-byte Ed25519 key, and derives the SUI address via
  BLAKE2b-256(0x00 || pubkey)
- `package.json`: adds `get-kms-address` script (`npm run get-kms-address`)
- `package.json`: adds `@aws-sdk/client-cloudformation` dependency used by
  the script to resolve the KMS key ID automatically
- `README.md`: new "Token Payments: Fund the KMS Hot Wallet" section with
  faucet links and instructions

## Dead CloudWatch metrics removed

The `SUIPayment` namespace metrics `InvoicesCreated` and `PaymentsDetected`
were defined in the CDK stack and referenced in a dashboard widget, but no
Lambda ever called `PutMetricData` to emit them — rendering the widget
permanently empty.

- Removed `invoiceCreatedMetric` and `paymentDetectedMetric` variables
- Replaced the "Invoice Lifecycle" dashboard widget (which referenced all
  four metrics including the two dead ones) with a "Sweep Results" widget
  showing only `SweepsSuccessful` and `SweepsFailed`, which the sweeper
  Lambda does actively emit
- The `cloudwatch:PutMetricData` IAM permission and the `failedSweepsAlarm`
  are retained — both are wired up to real sweeper behaviour

+++ non-evm-deployments/sui/invoice-generator/src/main.rs
@@ -24,7 +24,7 @@ fn is_token_whitelisted(token_address: &str) -> bool {
 
 #[derive(Deserialize)]
 struct CreateInvoiceRequest {
-    amount: u64,
+    amount: f64,
     reference_id: String,
     expiry_seconds: i64,
     #[serde(default)]
@@ -49,7 +49,7 @@ struct ApiGatewayResponse {
 struct CreateInvoiceResponse {
     invoice_id: String,
     recipient_address: String,
-    amount: u64,
+    amount: f64,
     expiry: i64,
     qr_code_base64: String,
 }
